requirements.txt
streamlit>=1.20
sumy>=0.9.0
newspaper3k>=0.2.8
nltk>=3.6
matplotlib>=3.5
beautifulsoup4>=4.9
wordcloud>=1.8
PyPDF2>=3.0
lxml_html_clean>=0.1
